chore: Split out `PickerGroupBox` to handle custom allocation behavior
